requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu116
torch == 1.12.1+cu116
torchvision == 0.13.1+cu116
torchaudio == 0.12.1+cu116

matplotlib == 3.6.2
numpy == 1.23.4
pandas == 1.5.1
Pillow == 9.2.0
seaborn == 0.12.1